+++ src/components/ViewContainer/ViewContainer.js
@@ -48,10 +48,7 @@ export const ViewContainer = memo(({
     const { reset: resetForm } = properties;
 
     resetForm();
-    transitionToParams({
-      _path: `${path}/view/${record.id}`,
-      layer: null,
-    });
+    history.push(`${path}/view/${record.id}`);
   };
 
   const deselectOnDelete = recordId => {
@@ -64,7 +61,7 @@ export const ViewContainer = memo(({
 
   const handleDeleteSuccess = record => {
     transitionToParams({
-      _path: `${path}/view`,
+      _path: `${path}`,
       layer: null,
     });
     deselectOnDelete(record.id);