+++ src/settings/JobProfiles/JobProfilesForm.js
@@ -99,7 +99,7 @@ export const JobProfilesFormComponent = memo(({
   stripes,
   parentResources,
   transitionToParams,
-  match: { path },
+  baseUrl,
   accordionStatusRef,
   layerType,
   mutator,
@@ -244,10 +244,7 @@ export const JobProfilesFormComponent = memo(({
         clearStorage();
 
         form.reset();
-        transitionToParams({
-          _path: `${path}/view/${record.id}`,
-          layer: null,
-        });
+        transitionToParams({ _path: `${baseUrl}/view/${record.id}` });
       }
     }
   };
@@ -390,7 +387,7 @@ JobProfilesFormComponent.propTypes = {
   stripes: PropTypes.object.isRequired,
   parentResources: PropTypes.object.isRequired,
   transitionToParams: PropTypes.func.isRequired,
-  match: PropTypes.shape({ path: PropTypes.string.isRequired }).isRequired,
+  baseUrl: PropTypes.string.isRequired,
   mutator: PropTypes.shape({ childWrappers: PropTypes.shape({ GET: PropTypes.func }) }).isRequired,
   resources: PropTypes.shape({ childWrappers: PropTypes.shape({ records: PropTypes.arrayOf(PropTypes.object) }) }).isRequired,
   accordionStatusRef: PropTypes.object,