+++ src/settings/JobProfiles/index.js
@@ -1,1 +1,4 @@
+export * from './CreateJobProfile';
+export * from './EditJobProfile';
+export * from './ViewJobProfile';
 export * from './JobProfiles';