+++ src/settings/MappingProfiles/detailsSections/view/tests/MappingMARCBibDetails.test.js
@@ -6,7 +6,10 @@ import '../../../../../../test/jest/__mock__';
 import { translationsProperties } from '../../../../../../test/jest/helpers';
 
 import { MappingMARCBibDetails } from '../MappingMARCBibDetails';
-import { FIELD_MAPPINGS_FOR_MARC } from '../../../../../utils';
+import {
+  FIELD_MAPPINGS_FOR_MARC,
+  MARC_TYPES,
+} from '../../../../../utils';
 
 const marcMappingDetailsProp = [];
 const mappingMarcFieldProtectionFieldsProp = [];
@@ -15,6 +18,7 @@ const marcFieldProtectionFieldsProp = [];
 const renderMappingMARCBibDetails = ({
   marcMappingOption,
   marcMappingDetails,
+  folioRecordType,
 }) => {
   const component = (
     <MappingMARCBibDetails
@@ -22,6 +26,7 @@ const renderMappingMARCBibDetails = ({
       marcMappingOption={marcMappingOption || FIELD_MAPPINGS_FOR_MARC.UPDATES}
       mappingMarcFieldProtectionFields={mappingMarcFieldProtectionFieldsProp}
       marcFieldProtectionFields={marcFieldProtectionFieldsProp}
+      folioRecordType={folioRecordType}
     />
   );
 
@@ -30,7 +35,7 @@ const renderMappingMARCBibDetails = ({
 
 describe('<MARCBibDetails>', () => {
   it('should have correct sections', () => {
-    const { getByRole } = renderMappingMARCBibDetails({});
+    const { getByRole } = renderMappingMARCBibDetails({ folioRecordType: MARC_TYPES.MARC_BIBLIOGRAPHIC });
 
     // eslint-disable-next-line no-irregular-whitespace
     expect(getByRole('button', { name: /icon field mapping · marc bibliographic · updates/i })).toBeInTheDocument();
@@ -39,14 +44,14 @@ describe('<MARCBibDetails>', () => {
   });
 
   it('"Updates" section is expanded by default', () => {
-    const { getByRole } = renderMappingMARCBibDetails({});
+    const { getByRole } = renderMappingMARCBibDetails({ folioRecordType: MARC_TYPES.MARC_BIBLIOGRAPHIC });
 
     // eslint-disable-next-line no-irregular-whitespace
     expect(getByRole('button', { name: /field mapping · marc bibliographic · updates/i })).toHaveAttribute('aria-expanded', 'true');
   });
 
   it('"Override protected fields" section is not expanded by default', () => {
-    const { getByRole } = renderMappingMARCBibDetails({});
+    const { getByRole } = renderMappingMARCBibDetails({ folioRecordType: MARC_TYPES.MARC_BIBLIOGRAPHIC });
 
     // eslint-disable-next-line no-irregular-whitespace
     expect(getByRole('button', { name: /field mapping · marc bibliographic · override protected fields/i }))
@@ -56,6 +61,7 @@ describe('<MARCBibDetails>', () => {
   describe('When there is mapping details', () => {
     it('should render contents as table', () => {
       const { getByText } = renderMappingMARCBibDetails({
+        folioRecordType: MARC_TYPES.MARC_BIBLIOGRAPHIC,
         marcMappingDetails: [{
           field: {
             field: 'testField',
@@ -83,6 +89,7 @@ describe('<MARCBibDetails>', () => {
     it('should render correct fields', () => {
       const { getByText } = renderMappingMARCBibDetails({
         marcMappingOption: FIELD_MAPPINGS_FOR_MARC.MODIFICATIONS,
+        folioRecordType: MARC_TYPES.MARC_BIBLIOGRAPHIC,
         marcMappingDetails: [{
           action: 'ADD',
           field: {