+++ src/settings/MappingProfiles/index.js
@@ -1,4 +1,6 @@
+export * from './CreateMappingProfile';
+export * from './EditMappingProfile';
+export * from './ViewMappingProfile';
 export * from './initialDetails';
 export * from './MappingProfiles';
-export * from './ViewMappingProfile';
 export * from './MappingProfilesForm';